app/pyproject.toml
[project]
name = "movie-finder-app"
version = "0.1.0"
description = "FastAPI application layer for Movie Finder"
license = { text = "MIT" }
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "movie-finder-chain",
    "imdbapi-client",
    "alembic>=1.17.1",
    "fastapi[standard]>=0.115.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "pydantic>=2.12.5",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.44",
]

[dependency-groups]
lint = ["ruff>=0.15.9", "mypy>=1.20.0"]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    "pre-commit>=4.5.1",
    "detect-secrets>=1.5.0",
]

[tool.uv]
default-groups = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
movie-finder-chain = { workspace = true }
imdbapi-client = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["chain"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
mypy_path = "src:tests"
strict = true
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
# jose stubs are incomplete; suppress misc errors on jwt calls
module = "app.auth.middleware"
disable_error_code = ["misc", "no-any-return"]

[[tool.mypy.overrides]]
# slowapi decorators are untyped; keep the suppression scoped to decorated route modules.
module = ["app.routers.auth", "app.routers.chat"]
disable_error_code = ["untyped-decorator"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "*/py.typed"]

[tool.coverage.report]
show_missing = true
